requirements.txt
2pandas
transformers
langchain
faiss-cpu
huggingface-hub
sentence-transformers
torch
